commit new

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -11,7 +11,7 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int xor_result = n ^ m;
+	unsigned long int xor_result = n ^ m;
 	unsigned int count = 0;
 
 	while (xor_result > 0)
